nvamemtiming: Handle vbios size bigger that 64K
Although vbios data is a little smaller that 64K, we may have dumped the whole PRAMIN size which is 128K (0x20000).
This is also better to remove this limitation as future vbios sizes may be bigger than 64K.

+++ include/nvamemtiming.h
@@ -58,7 +58,7 @@
 		struct {
 			const char *file;
 			uint8_t *data;
-			uint16_t length;
+			size_t length;
 
 			uint16_t timing_table_offset;
 			uint16_t timing_entry_offset;
